+++ lib/src/cook/process/toolchain/Configuration.cpp
@@ -4,13 +4,13 @@
 namespace cook { namespace process { namespace toolchain {
 
 
-    Configuration::Configuration(priority_type priority, const std::string & uuid) : 
+    ConfigurationCallback::ConfigurationCallback(Priority priority, const std::string & uuid) : 
         uuid(uuid),
         priority(priority)
     {
     }
 
-    bool Configuration::operator<(const Configuration & rhs) const
+    bool ConfigurationCallback::operator<(const ConfigurationCallback & rhs) const
     {
         return (priority < rhs.priority) ||
         (priority == rhs.priority && uuid < rhs.uuid);
@@ -27,78 +27,109 @@ namespace cook { namespace process { namespace toolchain {
         auto s = log::scope("Configuration value", [&](auto & n){
             n.attr("key", key).attr("value", value);
         });
-        config_.insert(std::make_pair(KeyValue(key, value), New));
+
+        const KeyValue kv{key, value};
+
+        for (const auto &[cb, state__kvs]: cb__state__kvs_)
+            for (const auto &[state, kvs]: state__kvs)
+                if (kvs.count(kv))
+                    return;
+
+        if (!cb__state__kvs_.empty())
+            cb__state__kvs_.begin()->second[New].insert(kv);
     }
 
-    bool ConfigurationBoard::add_callback(const Configuration & config)
+    bool ConfigurationBoard::add_callback(const ConfigurationCallback & config_callback)
     {
-        return config_cbs_.insert(config).second;
+        return cb__state__kvs_.emplace(config_callback, State__KeyValues{}).second;
     } 
 
     bool ConfigurationBoard::has_config(const std::string & key, const std::string & value) const
     {
-        return config_.find(KeyValue(key, value)) != config_.end();
+        const KeyValue needle{key, value};
+
+        for (const auto &[cb, state__kvs]: cb__state__kvs_)
+        {
+            for (const auto &[state, kvs]: state__kvs)
+            {
+                if (kvs.count(needle))
+                    return true;
+            }
+        }
+
+        return false;
     }
     bool ConfigurationBoard::has_config(const std::string & key) const
     {
-        auto lambda = [&](const auto & p)
+        for (const auto &[cb, state__kvs]: cb__state__kvs_)
         {
-            return p.first.first == key;
-        };
-        return std::find_if(config_.begin(), config_.end(), lambda) != config_.end();
+            for (const auto &[state, kvs]: state__kvs)
+            {
+                for (const auto &kv: kvs)
+                {
+                    if (kv.first == key)
+                        return true;
+                }
+            }
+        }
+
+        return false;
     }
 
     bool ConfigurationBoard::remove_config(const std::string & key)
     {
         unsigned int nr = 0;
-        for (auto it = config_.begin(); it != config_.end();)
+        for (auto &[cb, state__kvs]: cb__state__kvs_)
         {
-            if (it->first.first == key)
+            for (auto &[state, kvs]: state__kvs)
             {
-                it = config_.erase(it);
-                ++nr;
+                for (auto it = kvs.begin(); it != kvs.end();)
+                {
+                    if (it->first == key)
+                    {
+                        it = kvs.erase(it);
+                        ++nr;
+                    }
+                    else
+                        ++it;
+                }
             }
-            else
-                ++it;
         }
         return nr > 0;
     }
     bool ConfigurationBoard::remove_config(const std::string & key, const std::string & value)
     {
-        auto it = config_.find(KeyValue(key, value));
-        if (it == config_.end())
-            return false;
-        config_.erase(it);
-        return true;
+        const KeyValue needle{key, value};
+
+        unsigned int nr = 0;
+        for (auto &[cb, state__kvs]: cb__state__kvs_)
+        {
+            for (auto &[state, kvs]: state__kvs)
+            {
+                nr += kvs.erase(needle);
+            }
+        }
+
+        return nr > 0;
     }
 
     std::list<std::string> ConfigurationBoard::config_values(const std::string & key) const
     {
         std::list<std::string> result;
-        for(const auto & el : config_)
-            if (el.first.first == key)
-                result.push_back(el.first.second);
 
-        return result;
-    }
-
-    bool ConfigurationBoard::resolve_(Element::Ptr element, const KeyValue & kv)
-    {
-        auto s = log::scope("Resolving", [&](auto & n) {
-            n.attr("element_type", element->element_type()).attr("element_language", element->language()).attr("target_type", element->target_type()).attr("key", kv.first).attr("value", kv.second);
-        });
-        for(const Configuration & cfg : config_cbs_)
+        for (const auto &[cb, state__kvs]: cb__state__kvs_)
         {
-            const bool res = cfg.callback(element, kv.first, kv.second, *this);
-            auto ss = log::scope("callback", [&](auto & n) {
-                n.attr("uuid", cfg.uuid).attr("result", res);
-            });
-
-            if (res)
-                return true;
+            for (const auto &[state, kvs]: state__kvs)
+            {
+                for (const auto &kv: kvs)
+                {
+                    if (kv.first == key)
+                        result.push_back(kv.second);
+                }
+            }
         }
 
-        return false;
+        return result;
     }
 
 } } }